Restructured design items
