+++ lib/design/design.dart
@@ -1,0 +1,19 @@
+// It's better to ignore ordering to ease the grouping
+// The reason of this exports is when we are using design items, we only have one single source of import
+// ignore_for_file: directives_ordering
+
+// Base files
+export "src/colors.dart";
+export "src/display_sizes.dart";
+
+// Molecules
+export "src/molecule/info_item.dart";
+
+// Organisms
+export "src/organism/card_character.dart";
+export "src/organism/card_homeworld.dart";
+export "src/organism/card_starship.dart";
+export "src/organism/card_vehicle.dart";
+export "src/organism/loading.dart";
+export "src/text_sizes.dart";
+export "src/text_theme.dart";
